Lot's of repairs
Have the "Example.com (https:// will be added...) " Just show the "https://" and have it be a like darker.

The "+" icon is confusing next to the "Add RL" Change it, or even put that statement about being "Synced ot Cloud:..." btw which cloud is being synced to anyway?

The analitics aren't necessary so have that be a button at the bottom.

The date that the url was saved isn't necessary, take it out.
When you open up the qr code you only need the full url in centered and not in a white box - I thoght I could change it by the way it is displayed.

Whenever you give me new code, always give me the whole code content for me to copy and replace the old content.